Эта версия решает быстрее, но просит времени на иницилизацию (занимает 400 мегабайт оперативы). Вначале нужно 4 вызвать, потом уже работает гессер

+++ AutoGuesserTest/GuesserV1/AnswerVariantShortTest.cs
@@ -1,6 +1,6 @@
 using AutoGuesser;
-using NumberGameCore.BaseStuff;
-using NumberGameCore.BaseStuff.Holders;
+using GuesserDB.Entities;
+using GuesserDB.Entities.Holders;
 using NUnit.Framework;
 
 namespace AutoGuesserTest
@@ -13,7 +13,7 @@ namespace AutoGuesserTest
 			int[] expected = new[]{ 1, 2, 3, 4 };
 			Guess foo = SomeGuessHolder.GetByGuessed(expected);
 
-			int[] actual = foo.Guessed;
+			byte[] actual = foo.Guessed;
 			Assert.AreEqual(expected, actual);
 		}
 
@@ -27,10 +27,10 @@ namespace AutoGuesserTest
 
 			Guess foo = SomeGuessHolder.GetByGuessed(n1, n2, n3, n4);
 
-			int actual_n1 = foo.GetDigit(0);
-			int actual_n2 = foo.GetDigit(1);
-			int actual_n3 = foo.GetDigit(2);
-			int actual_n4 = foo.GetDigit(3);
+			byte actual_n1 = foo.GetDigit(0);
+			byte actual_n2 = foo.GetDigit(1);
+			byte actual_n3 = foo.GetDigit(2);
+			byte actual_n4 = foo.GetDigit(3);
 			Assert.AreEqual(n1, actual_n1);
 			Assert.AreEqual(n2, actual_n2);
 			Assert.AreEqual(n3, actual_n3);
